+++ Cargo.toml
@@ -1,9 +1,9 @@
 [package]
 
 name = "threadpool"
-version = "1.8.1"
+version = "2.0.0-alpha.1"
 edition = "2018"
-authors = ["The Rust Project Developers", "Corey Farwell <coreyf@rwell.org>", "Stefan Schindler <dns2utf8@estada.ch>"]
+authors = ["The Rust Project Developers", "Stefan Schindler <dns2utf8@estada.ch>"]
 license = "MIT/Apache-2.0"
 readme = "README.md"
 repository = "https://github.com/rust-threadpool/rust-threadpool"
@@ -24,4 +24,4 @@ include = [
 ]
 
 [dependencies]
-num_cpus = "1.13"
+num_cpus = "1.13.1"
