+++ src/lib.rs
@@ -20,12 +20,11 @@
 //! Every thread sends one message over the channel, which then is collected with the `take()`.
 //!
 //! ```
-//! use threadpool::ThreadPool;
 //! use std::sync::mpsc::channel;
 //!
 //! let n_workers = 4;
 //! let n_jobs = 8;
-//! let pool = ThreadPool::new(n_workers);
+//! let pool = threadpool::builder().num_workers(n_workers).build();
 //!
 //! let (tx, rx) = channel();
 //! for _ in 0..n_jobs {
@@ -34,6 +33,7 @@
 //!         tx.send(1).expect("channel will be there waiting for the pool");
 //!     });
 //! }
+//! drop(tx);
 //!
 //! assert_eq!(rx.iter().take(n_jobs).fold(0, |a, b| a + b), 8);
 //! ```
@@ -53,7 +53,7 @@
 //! // create at least as many workers as jobs or you will deadlock yourself
 //! let n_workers = 42;
 //! let n_jobs = 23;
-//! let pool = ThreadPool::new(n_workers);
+//! let pool = threadpool::builder().num_workers(n_workers).build();
 //! let an_atomic = Arc::new(AtomicUsize::new(0));
 //!
 //! assert!(n_jobs <= n_workers, "too many jobs, will deadlock");
@@ -75,7 +75,7 @@
 //!
 //! // wait for the threads to finish the work
 //! barrier.wait();
-//! assert_eq!(an_atomic.load(Ordering::SeqCst), /* n_jobs = */ 23);
+//! assert_eq!(an_atomic.load(Ordering::SeqCst), n_jobs);
 //! ```
 
 use num_cpus;
@@ -85,6 +85,70 @@ use std::sync::atomic::{AtomicUsize, Ordering};
 use std::sync::mpsc::{channel, Receiver, Sender};
 use std::sync::{Arc, Condvar, Mutex};
 use std::thread;
+
+#[cfg(test)]
+mod test;
+
+/// Creates a new thread pool with the same number of workers as CPUs are detected.
+///
+/// # Examples
+///
+/// Create a new thread pool capable of executing at least one jobs concurrently:
+///
+/// ```
+/// let pool = threadpool::auto_config();
+/// ```
+pub fn auto_config() -> ThreadPool {
+    builder().build()
+}
+/// Initiate a new [`Builder`].
+///
+/// [`Builder`]: struct.Builder.html
+///
+/// # Examples
+///
+/// ```
+/// let builder = threadpool::builder();
+/// ```
+pub const fn builder() -> Builder {
+    Builder {
+        num_workers: None,
+        worker_name: None,
+        thread_stack_size: None,
+    }
+}
+
+/*
+/// Creates a new thread pool capable of executing `num_workers` number of jobs concurrently.
+/// Each thread will have the [name][thread name] `name`.
+///
+/// # Panics
+///
+/// This function will panic if `num_workers` is 0.
+///
+/// # Examples
+///
+/// ```rust
+/// use std::thread;
+/// use threadpool::ThreadPool;
+///
+/// let pool = ThreadPool::with_name("worker".into(), 2);
+/// for _ in 0..2 {
+///     pool.execute(|| {
+///         assert_eq!(
+///             thread::current().name(),
+///             Some("worker")
+///         );
+///     });
+/// }
+/// pool.join();
+/// ```
+///
+/// [thread name]: https://doc.rust-lang.org/std/thread/struct.Thread.html#method.name
+pub fn with_name<S: AsRef<str>>(name: S) -> ThreadPool {
+    builder().worker_name(name).build()
+}
+*/
 
 trait FnBox {
     fn call_box(self: Box<Self>);
@@ -135,9 +199,9 @@ impl<'a> Drop for Sentinel<'a> {
 ///
 /// The three configuration options available:
 ///
-/// * `num_threads`: maximum number of threads that will be alive at any given moment by the built
+/// * `num_workers`: maximum number of threads that will be alive at any given moment by the built
 ///   [`ThreadPool`]
-/// * `thread_name`: thread name for each of the threads spawned by the built [`ThreadPool`]
+/// * `worker_name`: thread name for each of the threads spawned by the built [`ThreadPool`]
 /// * `thread_stack_size`: stack size (in bytes) for each of the threads spawned by the built
 ///   [`ThreadPool`]
 ///
@@ -149,36 +213,19 @@ impl<'a> Drop for Sentinel<'a> {
 /// a 8 MB stack size:
 ///
 /// ```
-/// let pool = threadpool::Builder::new()
-///     .num_threads(8)
-///     .thread_stack_size(8_000_000)
+/// let pool = threadpool::builder()
+///     .num_workers(8)
+///     .thread_stack_size(8 * 1024 * 1024)
 ///     .build();
 /// ```
 #[derive(Clone, Default)]
 pub struct Builder {
-    num_threads: Option<usize>,
-    thread_name: Option<String>,
+    num_workers: Option<usize>,
+    worker_name: Option<String>,
     thread_stack_size: Option<usize>,
 }
 
 impl Builder {
-    /// Initiate a new [`Builder`].
-    ///
-    /// [`Builder`]: struct.Builder.html
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// let builder = threadpool::Builder::new();
-    /// ```
-    pub fn new() -> Builder {
-        Builder {
-            num_threads: None,
-            thread_name: None,
-            thread_stack_size: None,
-        }
-    }
-
     /// Set the maximum number of worker-threads that will be alive at any given moment by the built
     /// [`ThreadPool`]. If not specified, defaults the number of threads to the number of CPUs.
     ///
@@ -186,7 +233,7 @@ impl Builder {
     ///
     /// # Panics
     ///
-    /// This method will panic if `num_threads` is 0.
+    /// This method will panic if `num_workers` is 0.
     ///
     /// # Examples
     ///
@@ -195,19 +242,19 @@ impl Builder {
     /// ```
     /// use std::thread;
     ///
-    /// let pool = threadpool::Builder::new()
-    ///     .num_threads(8)
+    /// let pool = threadpool::builder()
+    ///     .num_workers(8)
     ///     .build();
     ///
-    /// for _ in 0..100 {
+    /// for _ in 0..42 {
     ///     pool.execute(|| {
     ///         println!("Hello from a worker thread!")
     ///     })
     /// }
     /// ```
-    pub fn num_threads(mut self, num_threads: usize) -> Builder {
-        assert!(num_threads > 0);
-        self.num_threads = Some(num_threads);
+    pub fn num_workers(mut self, num_workers: usize) -> Builder {
+        assert!(num_workers > 0);
+        self.num_workers = Some(num_workers);
         self
     }
 
@@ -223,8 +270,8 @@ impl Builder {
     /// ```
     /// use std::thread;
     ///
-    /// let pool = threadpool::Builder::new()
-    ///     .thread_name("foo".into())
+    /// let pool = threadpool::builder()
+    ///     .worker_name("foo")
     ///     .build();
     ///
     /// for _ in 0..100 {
@@ -233,8 +280,9 @@ impl Builder {
     ///     })
     /// }
     /// ```
-    pub fn thread_name(mut self, name: String) -> Builder {
-        self.thread_name = Some(name);
+    pub fn worker_name<S: AsRef<str>>(mut self, name: S) -> Builder {
+        // TODO save the copy with Into<String>
+        self.worker_name = Some(name.as_ref().to_owned());
         self
     }
 
@@ -250,8 +298,8 @@ impl Builder {
     /// Each thread spawned by this pool will have a 4 MB stack:
     ///
     /// ```
-    /// let pool = threadpool::Builder::new()
-    ///     .thread_stack_size(4_000_000)
+    /// let pool = threadpool::builder()
+    ///     .thread_stack_size(4096 * 1024)
     ///     .build();
     ///
     /// for _ in 0..100 {
@@ -273,31 +321,31 @@ impl Builder {
     /// # Examples
     ///
     /// ```
-    /// let pool = threadpool::Builder::new()
-    ///     .num_threads(8)
-    ///     .thread_stack_size(4_000_000)
+    /// let pool = threadpool::builder()
+    ///     .num_workers(8)
+    ///     .thread_stack_size(16*1024*1024)
     ///     .build();
     /// ```
     pub fn build(self) -> ThreadPool {
         let (tx, rx) = channel::<Thunk<'static>>();
 
-        let num_threads = self.num_threads.unwrap_or_else(num_cpus::get);
+        let num_workers = self.num_workers.unwrap_or_else(num_cpus::get);
 
         let shared_data = Arc::new(ThreadPoolSharedData {
-            name: self.thread_name,
+            name: self.worker_name,
             job_receiver: Mutex::new(rx),
             empty_condvar: Condvar::new(),
             empty_trigger: Mutex::new(()),
             join_generation: AtomicUsize::new(0),
             queued_count: AtomicUsize::new(0),
             active_count: AtomicUsize::new(0),
-            max_thread_count: AtomicUsize::new(num_threads),
+            max_thread_count: AtomicUsize::new(num_workers),
             panic_count: AtomicUsize::new(0),
             stack_size: self.thread_stack_size,
         });
 
         // Threadpool threads
-        for _ in 0..num_threads {
+        for _ in 0..num_workers {
             spawn_in_pool(shared_data.clone());
         }
 
@@ -349,65 +397,6 @@ pub struct ThreadPool {
 }
 
 impl ThreadPool {
-    /// Creates a new thread pool capable of executing `num_threads` number of jobs concurrently.
-    ///
-    /// # Panics
-    ///
-    /// This function will panic if `num_threads` is 0.
-    ///
-    /// # Examples
-    ///
-    /// Create a new thread pool capable of executing four jobs concurrently:
-    ///
-    /// ```
-    /// use threadpool::ThreadPool;
-    ///
-    /// let pool = ThreadPool::new(4);
-    /// ```
-    pub fn new(num_threads: usize) -> ThreadPool {
-        Builder::new().num_threads(num_threads).build()
-    }
-
-    /// Creates a new thread pool capable of executing `num_threads` number of jobs concurrently.
-    /// Each thread will have the [name][thread name] `name`.
-    ///
-    /// # Panics
-    ///
-    /// This function will panic if `num_threads` is 0.
-    ///
-    /// # Examples
-    ///
-    /// ```rust
-    /// use std::thread;
-    /// use threadpool::ThreadPool;
-    ///
-    /// let pool = ThreadPool::with_name("worker".into(), 2);
-    /// for _ in 0..2 {
-    ///     pool.execute(|| {
-    ///         assert_eq!(
-    ///             thread::current().name(),
-    ///             Some("worker")
-    ///         );
-    ///     });
-    /// }
-    /// pool.join();
-    /// ```
-    ///
-    /// [thread name]: https://doc.rust-lang.org/std/thread/struct.Thread.html#method.name
-    pub fn with_name(name: String, num_threads: usize) -> ThreadPool {
-        Builder::new()
-            .num_threads(num_threads)
-            .thread_name(name)
-            .build()
-    }
-
-    /// **Deprecated: Use [`ThreadPool::with_name`](#method.with_name)**
-    #[inline(always)]
-    #[deprecated(since = "1.4.0", note = "use ThreadPool::with_name")]
-    pub fn new_with_name(name: String, num_threads: usize) -> ThreadPool {
-        Self::with_name(name, num_threads)
-    }
-
     /// Executes the function `job` on a thread in the pool.
     ///
     /// # Examples
@@ -415,9 +404,7 @@ impl ThreadPool {
     /// Execute four jobs on a thread pool that can run two jobs concurrently:
     ///
     /// ```
-    /// use threadpool::ThreadPool;
-    ///
-    /// let pool = ThreadPool::new(2);
+    /// let pool = threadpool::auto_config();
     /// pool.execute(|| println!("hello"));
     /// pool.execute(|| println!("world"));
     /// pool.execute(|| println!("foo"));
@@ -443,7 +430,7 @@ impl ThreadPool {
     /// use std::time::Duration;
     /// use std::thread::sleep;
     ///
-    /// let pool = ThreadPool::new(2);
+    /// let pool = threadpool::builder().num_workers(2).build();
     /// for _ in 0..10 {
     ///     pool.execute(|| {
     ///         sleep(Duration::from_secs(100));
@@ -457,16 +444,15 @@ impl ThreadPool {
         self.shared_data.queued_count.load(Ordering::Relaxed)
     }
 
-    /// Returns the number of currently active threads.
+    /// Returns the number of currently active worker threads.
     ///
     /// # Examples
     ///
     /// ```
-    /// use threadpool::ThreadPool;
     /// use std::time::Duration;
     /// use std::thread::sleep;
     ///
-    /// let pool = ThreadPool::new(4);
+    /// let pool = threadpool::builder().num_workers(4).build();
     /// for _ in 0..10 {
     ///     pool.execute(move || {
     ///         sleep(Duration::from_secs(100));
@@ -485,12 +471,10 @@ impl ThreadPool {
     /// # Examples
     ///
     /// ```
-    /// use threadpool::ThreadPool;
-    ///
-    /// let mut pool = ThreadPool::new(4);
+    /// let pool = threadpool::builder().num_workers(4).build();
     /// assert_eq!(4, pool.max_count());
     ///
-    /// pool.set_num_threads(8);
+    /// pool.set_num_workers(8);
     /// assert_eq!(8, pool.max_count());
     /// ```
     pub fn max_count(&self) -> usize {
@@ -502,9 +486,7 @@ impl ThreadPool {
     /// # Examples
     ///
     /// ```
-    /// use threadpool::ThreadPool;
-    ///
-    /// let pool = ThreadPool::new(4);
+    /// let pool = threadpool::auto_config();
     /// for n in 0..10 {
     ///     pool.execute(move || {
     ///         // simulate a panic
@@ -521,28 +503,21 @@ impl ThreadPool {
         self.shared_data.panic_count.load(Ordering::Relaxed)
     }
 
-    /// **Deprecated: Use [`ThreadPool::set_num_threads`](#method.set_num_threads)**
-    #[deprecated(since = "1.3.0", note = "use ThreadPool::set_num_threads")]
-    pub fn set_threads(&mut self, num_threads: usize) {
-        self.set_num_threads(num_threads)
-    }
-
-    /// Sets the number of worker-threads to use as `num_threads`.
+    /// Sets the number of worker-threads to use as `num_workers`.
     /// Can be used to change the threadpool size during runtime.
     /// Will not abort already running or waiting threads.
     ///
     /// # Panics
     ///
-    /// This function will panic if `num_threads` is 0.
+    /// This function will panic if `num_workers` is 0.
     ///
     /// # Examples
     ///
     /// ```
-    /// use threadpool::ThreadPool;
     /// use std::time::Duration;
     /// use std::thread::sleep;
     ///
-    /// let mut pool = ThreadPool::new(4);
+    /// let  pool = threadpool::builder().num_workers(4).build();
     /// for _ in 0..10 {
     ///     pool.execute(move || {
     ///         sleep(Duration::from_secs(100));
@@ -554,7 +529,7 @@ impl ThreadPool {
     /// assert_eq!(6, pool.queued_count());
     ///
     /// // Increase thread capacity of the pool
-    /// pool.set_num_threads(8);
+    /// pool.set_num_workers(8);
     ///
     /// sleep(Duration::from_secs(1)); // wait for new threads to start
     /// assert_eq!(8, pool.active_count());
@@ -562,18 +537,18 @@ impl ThreadPool {
     ///
     /// // Decrease thread capacity of the pool
     /// // No active threads are killed
-    /// pool.set_num_threads(4);
+    /// pool.set_num_workers(4);
     ///
     /// assert_eq!(8, pool.active_count());
     /// assert_eq!(2, pool.queued_count());
     /// ```
-    pub fn set_num_threads(&mut self, num_threads: usize) {
-        assert!(num_threads >= 1);
-        let prev_num_threads = self
+    pub fn set_num_workers(&self, num_workers: usize) {
+        assert!(num_workers >= 1);
+        let prev_num_workers = self
             .shared_data
             .max_thread_count
-            .swap(num_threads, Ordering::Release);
-        if let Some(num_spawn) = num_threads.checked_sub(prev_num_threads) {
+            .swap(num_workers, Ordering::Release);
+        if let Some(num_spawn) = num_workers.checked_sub(prev_num_workers) {
             // Spawn new threads
             for _ in 0..num_spawn {
                 spawn_in_pool(self.shared_data.clone());
@@ -592,6 +567,9 @@ impl ThreadPool {
     /// Calling `join` from a thread within the pool will cause a deadlock. This
     /// behavior is considered safe.
     ///
+    /// **Note:** Join will not stop the worker threads. You will need to `drop`
+    /// all instances of `ThreadPool` for the worker threads to terminate.
+    ///
     /// # Examples
     ///
     /// ```
@@ -599,7 +577,7 @@ impl ThreadPool {
     /// use std::sync::Arc;
     /// use std::sync::atomic::{AtomicUsize, Ordering};
     ///
-    /// let pool = ThreadPool::new(8);
+    /// let pool = threadpool::auto_config();
     /// let test_count = Arc::new(AtomicUsize::new(0));
     ///
     /// for _ in 0..42 {
@@ -627,10 +605,11 @@ impl ThreadPool {
             lock = self.shared_data.empty_condvar.wait(lock).unwrap();
         }
 
-        // increase generation if we are the first thread to come out of the loop
-        self.shared_data.join_generation.compare_and_swap(
+        // increase generation if we are the first joining thread to come out of the loop
+        let _ = self.shared_data.join_generation.compare_exchange(
             generation,
             generation.wrapping_add(1),
+            Ordering::SeqCst,
             Ordering::SeqCst,
         );
     }
@@ -643,11 +622,10 @@ impl Clone for ThreadPool {
     /// We could for example submit jobs from multiple threads concurrently.
     ///
     /// ```
-    /// use threadpool::ThreadPool;
     /// use std::thread;
     /// use std::sync::mpsc::channel;
     ///
-    /// let pool = ThreadPool::with_name("clone example".into(), 2);
+    /// let pool = threadpool::builder().worker_name("clone example").num_workers(2).build();
     ///
     /// let results = (0..2)
     ///     .map(|i| {
@@ -681,15 +659,6 @@ impl Clone for ThreadPool {
     }
 }
 
-/// Create a thread pool with one thread per CPU.
-/// On machines with hyperthreading,
-/// this will create one thread per hyperthread.
-impl Default for ThreadPool {
-    fn default() -> Self {
-        ThreadPool::new(num_cpus::get())
-    }
-}
-
 impl fmt::Debug for ThreadPool {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         f.debug_struct("ThreadPool")
@@ -705,24 +674,17 @@ impl PartialEq for ThreadPool {
     /// Check if you are working with the same pool
     ///
     /// ```
-    /// use threadpool::ThreadPool;
-    ///
-    /// let a = ThreadPool::new(2);
-    /// let b = ThreadPool::new(2);
+    /// let a = threadpool::auto_config();
+    /// let b = threadpool::auto_config();
     ///
     /// assert_eq!(a, a);
     /// assert_eq!(b, b);
     ///
-    /// # // TODO: change this to assert_ne in the future
-    /// assert!(a != b);
-    /// assert!(b != a);
+    /// assert_ne!(a, b);
+    /// assert_ne!(b, a);
     /// ```
     fn eq(&self, other: &ThreadPool) -> bool {
-        let a: &ThreadPoolSharedData = &*self.shared_data;
-        let b: &ThreadPoolSharedData = &*other.shared_data;
-        a as *const ThreadPoolSharedData == b as *const ThreadPoolSharedData
-        // with rust 1.17 and late:
-        // Arc::ptr_eq(&self.shared_data, &other.shared_data)
+        Arc::ptr_eq(&self.shared_data, &other.shared_data)
     }
 }
 impl Eq for ThreadPool {}
@@ -775,555 +737,4 @@ fn spawn_in_pool(shared_data: Arc<ThreadPoolSharedData>) {
             sentinel.cancel();
         })
         .unwrap();
-}
-
-#[cfg(test)]
-mod test {
-    use super::{Builder, ThreadPool};
-    use std::sync::atomic::{AtomicUsize, Ordering};
-    use std::sync::mpsc::{channel, sync_channel};
-    use std::sync::{Arc, Barrier};
-    use std::thread::{self, sleep};
-    use std::time::Duration;
-
-    const TEST_TASKS: usize = 4;
-
-    #[test]
-    fn test_set_num_threads_increasing() {
-        let new_thread_amount = TEST_TASKS + 8;
-        let mut pool = ThreadPool::new(TEST_TASKS);
-        for _ in 0..TEST_TASKS {
-            pool.execute(move || sleep(Duration::from_secs(23)));
-        }
-        sleep(Duration::from_secs(1));
-        assert_eq!(pool.active_count(), TEST_TASKS);
-
-        pool.set_num_threads(new_thread_amount);
-
-        for _ in 0..(new_thread_amount - TEST_TASKS) {
-            pool.execute(move || sleep(Duration::from_secs(23)));
-        }
-        sleep(Duration::from_secs(1));
-        assert_eq!(pool.active_count(), new_thread_amount);
-
-        pool.join();
-    }
-
-    #[test]
-    fn test_set_num_threads_decreasing() {
-        let new_thread_amount = 2;
-        let mut pool = ThreadPool::new(TEST_TASKS);
-        for _ in 0..TEST_TASKS {
-            pool.execute(move || {
-                assert_eq!(1, 1);
-            });
-        }
-        pool.set_num_threads(new_thread_amount);
-        for _ in 0..new_thread_amount {
-            pool.execute(move || sleep(Duration::from_secs(23)));
-        }
-        sleep(Duration::from_secs(1));
-        assert_eq!(pool.active_count(), new_thread_amount);
-
-        pool.join();
-    }
-
-    #[test]
-    fn test_active_count() {
-        let pool = ThreadPool::new(TEST_TASKS);
-        for _ in 0..2 * TEST_TASKS {
-            pool.execute(move || loop {
-                sleep(Duration::from_secs(10))
-            });
-        }
-        sleep(Duration::from_secs(1));
-        let active_count = pool.active_count();
-        assert_eq!(active_count, TEST_TASKS);
-        let initialized_count = pool.max_count();
-        assert_eq!(initialized_count, TEST_TASKS);
-    }
-
-    #[test]
-    fn test_works() {
-        let pool = ThreadPool::new(TEST_TASKS);
-
-        let (tx, rx) = channel();
-        for _ in 0..TEST_TASKS {
-            let tx = tx.clone();
-            pool.execute(move || {
-                tx.send(1).unwrap();
-            });
-        }
-
-        assert_eq!(rx.iter().take(TEST_TASKS).fold(0, |a, b| a + b), TEST_TASKS);
-    }
-
-    #[test]
-    #[should_panic]
-    fn test_zero_tasks_panic() {
-        ThreadPool::new(0);
-    }
-
-    #[test]
-    fn test_recovery_from_subtask_panic() {
-        let pool = ThreadPool::new(TEST_TASKS);
-
-        // Panic all the existing threads.
-        for _ in 0..TEST_TASKS {
-            pool.execute(move || panic!("Ignore this panic, it must!"));
-        }
-        pool.join();
-
-        assert_eq!(pool.panic_count(), TEST_TASKS);
-
-        // Ensure new threads were spawned to compensate.
-        let (tx, rx) = channel();
-        for _ in 0..TEST_TASKS {
-            let tx = tx.clone();
-            pool.execute(move || {
-                tx.send(1).unwrap();
-            });
-        }
-
-        assert_eq!(rx.iter().take(TEST_TASKS).fold(0, |a, b| a + b), TEST_TASKS);
-    }
-
-    #[test]
-    fn test_should_not_panic_on_drop_if_subtasks_panic_after_drop() {
-        let pool = ThreadPool::new(TEST_TASKS);
-        let waiter = Arc::new(Barrier::new(TEST_TASKS + 1));
-
-        // Panic all the existing threads in a bit.
-        for _ in 0..TEST_TASKS {
-            let waiter = waiter.clone();
-            pool.execute(move || {
-                waiter.wait();
-                panic!("Ignore this panic, it should!");
-            });
-        }
-
-        drop(pool);
-
-        // Kick off the failure.
-        waiter.wait();
-    }
-
-    #[test]
-    fn test_massive_task_creation() {
-        let test_tasks = 4_200_000;
-
-        let pool = ThreadPool::new(TEST_TASKS);
-        let b0 = Arc::new(Barrier::new(TEST_TASKS + 1));
-        let b1 = Arc::new(Barrier::new(TEST_TASKS + 1));
-
-        let (tx, rx) = channel();
-
-        for i in 0..test_tasks {
-            let tx = tx.clone();
-            let (b0, b1) = (b0.clone(), b1.clone());
-
-            pool.execute(move || {
-                // Wait until the pool has been filled once.
-                if i < TEST_TASKS {
-                    b0.wait();
-                    // wait so the pool can be measured
-                    b1.wait();
-                }
-
-                let _ = tx.send(1);
-            });
-        }
-
-        b0.wait();
-        assert_eq!(pool.active_count(), TEST_TASKS);
-        b1.wait();
-
-        assert_eq!(rx.iter().take(test_tasks).fold(0, |a, b| a + b), test_tasks);
-        pool.join();
-
-        let atomic_active_count = pool.active_count();
-        assert!(
-            atomic_active_count == 0,
-            "atomic_active_count: {}",
-            atomic_active_count
-        );
-    }
-
-    #[test]
-    fn test_shrink() {
-        let test_tasks_begin = TEST_TASKS + 2;
-
-        let mut pool = ThreadPool::new(test_tasks_begin);
-        let b0 = Arc::new(Barrier::new(test_tasks_begin + 1));
-        let b1 = Arc::new(Barrier::new(test_tasks_begin + 1));
-
-        for _ in 0..test_tasks_begin {
-            let (b0, b1) = (b0.clone(), b1.clone());
-            pool.execute(move || {
-                b0.wait();
-                b1.wait();
-            });
-        }
-
-        let b2 = Arc::new(Barrier::new(TEST_TASKS + 1));
-        let b3 = Arc::new(Barrier::new(TEST_TASKS + 1));
-
-        for _ in 0..TEST_TASKS {
-            let (b2, b3) = (b2.clone(), b3.clone());
-            pool.execute(move || {
-                b2.wait();
-                b3.wait();
-            });
-        }
-
-        b0.wait();
-        pool.set_num_threads(TEST_TASKS);
-
-        assert_eq!(pool.active_count(), test_tasks_begin);
-        b1.wait();
-
-        b2.wait();
-        assert_eq!(pool.active_count(), TEST_TASKS);
-        b3.wait();
-    }
-
-    #[test]
-    fn test_name() {
-        let name = "test";
-        let mut pool = ThreadPool::with_name(name.to_owned(), 2);
-        let (tx, rx) = sync_channel(0);
-
-        // initial thread should share the name "test"
-        for _ in 0..2 {
-            let tx = tx.clone();
-            pool.execute(move || {
-                let name = thread::current().name().unwrap().to_owned();
-                tx.send(name).unwrap();
-            });
-        }
-
-        // new spawn thread should share the name "test" too.
-        pool.set_num_threads(3);
-        let tx_clone = tx.clone();
-        pool.execute(move || {
-            let name = thread::current().name().unwrap().to_owned();
-            tx_clone.send(name).unwrap();
-            panic!();
-        });
-
-        // recover thread should share the name "test" too.
-        pool.execute(move || {
-            let name = thread::current().name().unwrap().to_owned();
-            tx.send(name).unwrap();
-        });
-
-        for thread_name in rx.iter().take(4) {
-            assert_eq!(name, thread_name);
-        }
-    }
-
-    #[test]
-    fn test_debug() {
-        let pool = ThreadPool::new(4);
-        let debug = format!("{:?}", pool);
-        assert_eq!(
-            debug,
-            "ThreadPool { name: None, queued_count: 0, active_count: 0, max_count: 4 }"
-        );
-
-        let pool = ThreadPool::with_name("hello".into(), 4);
-        let debug = format!("{:?}", pool);
-        assert_eq!(
-            debug,
-            "ThreadPool { name: Some(\"hello\"), queued_count: 0, active_count: 0, max_count: 4 }"
-        );
-
-        let pool = ThreadPool::new(4);
-        pool.execute(move || sleep(Duration::from_secs(5)));
-        sleep(Duration::from_secs(1));
-        let debug = format!("{:?}", pool);
-        assert_eq!(
-            debug,
-            "ThreadPool { name: None, queued_count: 0, active_count: 1, max_count: 4 }"
-        );
-    }
-
-    #[test]
-    fn test_repeate_join() {
-        let pool = ThreadPool::with_name("repeate join test".into(), 8);
-        let test_count = Arc::new(AtomicUsize::new(0));
-
-        for _ in 0..42 {
-            let test_count = test_count.clone();
-            pool.execute(move || {
-                sleep(Duration::from_secs(2));
-                test_count.fetch_add(1, Ordering::Release);
-            });
-        }
-
-        println!("{:?}", pool);
-        pool.join();
-        assert_eq!(42, test_count.load(Ordering::Acquire));
-
-        for _ in 0..42 {
-            let test_count = test_count.clone();
-            pool.execute(move || {
-                sleep(Duration::from_secs(2));
-                test_count.fetch_add(1, Ordering::Relaxed);
-            });
-        }
-        pool.join();
-        assert_eq!(84, test_count.load(Ordering::Relaxed));
-    }
-
-    #[test]
-    fn test_multi_join() {
-        use std::sync::mpsc::TryRecvError::*;
-
-        // Toggle the following lines to debug the deadlock
-        fn error(_s: String) {
-            //use ::std::io::Write;
-            //let stderr = ::std::io::stderr();
-            //let mut stderr = stderr.lock();
-            //stderr.write(&_s.as_bytes()).is_ok();
-        }
-
-        let pool0 = ThreadPool::with_name("multi join pool0".into(), 4);
-        let pool1 = ThreadPool::with_name("multi join pool1".into(), 4);
-        let (tx, rx) = channel();
-
-        for i in 0..8 {
-            let pool1 = pool1.clone();
-            let pool0_ = pool0.clone();
-            let tx = tx.clone();
-            pool0.execute(move || {
-                pool1.execute(move || {
-                    error(format!("p1: {} -=- {:?}\n", i, pool0_));
-                    pool0_.join();
-                    error(format!("p1: send({})\n", i));
-                    tx.send(i).expect("send i from pool1 -> main");
-                });
-                error(format!("p0: {}\n", i));
-            });
-        }
-        drop(tx);
-
-        assert_eq!(rx.try_recv(), Err(Empty));
-        error(format!("{:?}\n{:?}\n", pool0, pool1));
-        pool0.join();
-        error(format!("pool0.join() complete =-= {:?}", pool1));
-        pool1.join();
-        error("pool1.join() complete\n".into());
-        assert_eq!(
-            rx.iter().fold(0, |acc, i| acc + i),
-            0 + 1 + 2 + 3 + 4 + 5 + 6 + 7
-        );
-    }
-
-    #[test]
-    fn test_empty_pool() {
-        // Joining an empty pool must return imminently
-        let pool = ThreadPool::new(4);
-
-        pool.join();
-
-        assert!(true);
-    }
-
-    #[test]
-    fn test_no_fun_or_joy() {
-        // What happens when you keep adding jobs after a join
-
-        fn sleepy_function() {
-            sleep(Duration::from_secs(6));
-        }
-
-        let pool = ThreadPool::with_name("no fun or joy".into(), 8);
-
-        pool.execute(sleepy_function);
-
-        let p_t = pool.clone();
-        thread::spawn(move || {
-            (0..23).map(|_| p_t.execute(sleepy_function)).count();
-        });
-
-        pool.join();
-    }
-
-    #[test]
-    fn test_clone() {
-        let pool = ThreadPool::with_name("clone example".into(), 2);
-
-        // This batch of jobs will occupy the pool for some time
-        for _ in 0..6 {
-            pool.execute(move || {
-                sleep(Duration::from_secs(2));
-            });
-        }
-
-        // The following jobs will be inserted into the pool in a random fashion
-        let t0 = {
-            let pool = pool.clone();
-            thread::spawn(move || {
-                // wait for the first batch of tasks to finish
-                pool.join();
-
-                let (tx, rx) = channel();
-                for i in 0..42 {
-                    let tx = tx.clone();
-                    pool.execute(move || {
-                        tx.send(i).expect("channel will be waiting");
-                    });
-                }
-                drop(tx);
-                rx.iter()
-                    .fold(0, |accumulator, element| accumulator + element)
-            })
-        };
-        let t1 = {
-            let pool = pool.clone();
-            thread::spawn(move || {
-                // wait for the first batch of tasks to finish
-                pool.join();
-
-                let (tx, rx) = channel();
-                for i in 1..12 {
-                    let tx = tx.clone();
-                    pool.execute(move || {
-                        tx.send(i).expect("channel will be waiting");
-                    });
-                }
-                drop(tx);
-                rx.iter()
-                    .fold(1, |accumulator, element| accumulator * element)
-            })
-        };
-
-        assert_eq!(
-            861,
-            t0.join()
-                .expect("thread 0 will return after calculating additions",)
-        );
-        assert_eq!(
-            39916800,
-            t1.join()
-                .expect("thread 1 will return after calculating multiplications",)
-        );
-    }
-
-    #[test]
-    fn test_sync_shared_data() {
-        fn assert_sync<T: Sync>() {}
-        assert_sync::<super::ThreadPoolSharedData>();
-    }
-
-    #[test]
-    fn test_send_shared_data() {
-        fn assert_send<T: Send>() {}
-        assert_send::<super::ThreadPoolSharedData>();
-    }
-
-    #[test]
-    fn test_send() {
-        fn assert_send<T: Send>() {}
-        assert_send::<ThreadPool>();
-    }
-
-    #[test]
-    fn test_cloned_eq() {
-        let a = ThreadPool::new(2);
-
-        assert_eq!(a, a.clone());
-    }
-
-    #[test]
-    /// The scenario is joining threads should not be stuck once their wave
-    /// of joins has completed. So once one thread joining on a pool has
-    /// succeded other threads joining on the same pool must get out even if
-    /// the thread is used for other jobs while the first group is finishing
-    /// their join
-    ///
-    /// In this example this means the waiting threads will exit the join in
-    /// groups of four because the waiter pool has four workers.
-    fn test_join_wavesurfer() {
-        let n_cycles = 4;
-        let n_workers = 4;
-        let (tx, rx) = channel();
-        let builder = Builder::new()
-            .num_threads(n_workers)
-            .thread_name("join wavesurfer".into());
-        let p_waiter = builder.clone().build();
-        let p_clock = builder.build();
-
-        let barrier = Arc::new(Barrier::new(3));
-        let wave_clock = Arc::new(AtomicUsize::new(0));
-        let clock_thread = {
-            let barrier = barrier.clone();
-            let wave_clock = wave_clock.clone();
-            thread::spawn(move || {
-                barrier.wait();
-                for wave_num in 0..n_cycles {
-                    wave_clock.store(wave_num, Ordering::SeqCst);
-                    sleep(Duration::from_secs(1));
-                }
-            })
-        };
-
-        {
-            let barrier = barrier.clone();
-            p_clock.execute(move || {
-                barrier.wait();
-                // this sleep is for stabilisation on weaker platforms
-                sleep(Duration::from_millis(100));
-            });
-        }
-
-        // prepare three waves of jobs
-        for i in 0..3 * n_workers {
-            let p_clock = p_clock.clone();
-            let tx = tx.clone();
-            let wave_clock = wave_clock.clone();
-            p_waiter.execute(move || {
-                let now = wave_clock.load(Ordering::SeqCst);
-                p_clock.join();
-                // submit jobs for the second wave
-                p_clock.execute(|| sleep(Duration::from_secs(1)));
-                let clock = wave_clock.load(Ordering::SeqCst);
-                tx.send((now, clock, i)).unwrap();
-            });
-        }
-        println!("all scheduled at {}", wave_clock.load(Ordering::SeqCst));
-        barrier.wait();
-
-        p_clock.join();
-        //p_waiter.join();
-
-        drop(tx);
-        let mut hist = vec![0; n_cycles];
-        let mut data = vec![];
-        for (now, after, i) in rx.iter() {
-            let mut dur = after - now;
-            if dur >= n_cycles - 1 {
-                dur = n_cycles - 1;
-            }
-            hist[dur] += 1;
-
-            data.push((now, after, i));
-        }
-        for (i, n) in hist.iter().enumerate() {
-            println!(
-                "\t{}: {} {}",
-                i,
-                n,
-                &*(0..*n).fold("".to_owned(), |s, _| s + "*")
-            );
-        }
-        assert!(data.iter().all(|&(cycle, stop, i)| if i < n_workers {
-            cycle == stop
-        } else {
-            cycle < stop
-        }));
-
-        clock_thread.join().unwrap();
-    }
 }
